dev-requirements.txt
requests
beautifulsoup4
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
alembic
psycopg2